+++ changelog.md
@@ -9,6 +9,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 * `setup-main-chain-state` command now uses native Rust to upsert the D-Parameter and upsert permissioned candidates
 * Changed the `smart-contracts init-governance` command to `smart-contracts governance init`
 * Update dependencies containers to cardano-node 10.1.4, db-sync 13.6.0.4, ogmios 6.11.0 and kupo 2.10.0
+* smart-contracts commands and offchain tests now use WebSockets implementation of Ogmios client
 
 ## Removed
 
@@ -25,6 +26,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 * Command `smart-contracts reserve init`
 * Command `smart-contracts reserve create`
 * Command `smart-contracts governance update`
+* Ogmios client backed by jsonrpsee `WsClient`
 
 # v1.4.0